+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/DownloadsFeature.kt
@@ -15,25 +15,28 @@ import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.feature.downloads.DownloadDialogFragment.Companion.FRAGMENT_TAG
 import mozilla.components.support.base.feature.LifecycleAwareFeature
+import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.ktx.android.content.isPermissionGranted
 
 typealias OnNeedToRequestPermissions = (session: Session, download: Download) -> Unit
 
 /**
- * Feature implementation for proving download functionality for the selected session.
- * It will subscribe to the selected session and will listening for downloads.
+ * Feature implementation to provide download functionality for the selected
+ * session. The feature will subscribe to the selected session and listen
+ * for downloads.
  *
- * @property applicationContext a reference to [Context] applicationContext.
- * @property onNeedToRequestPermissions it will be call when you need to request permission,
- * before a download can be performed.
- * @property onDownloadCompleted a callback to be notified when a download is completed.
- * @property downloadManager a reference of [DownloadManager] who is in charge of performing the downloads.
- * @property sessionManager a reference of [SessionManager].
- * @property fragmentManager a reference of [FragmentManager]. If you provide a [fragmentManager]
- * a dialog will show before every download.
- * @property fragmentManager a reference of [FragmentManager].
- * @property dialog a reference of [DownloadDialogFragment]. If it is not provided, it will be
- * initialized with an instance of [SimpleDownloadDialogFragment].
+ * @property applicationContext a reference to the application context.
+ * @property onNeedToRequestPermissions a callback invoked when permissions
+ * need to be requested before a download can be performed. At a minimum,
+ * [WRITE_EXTERNAL_STORAGE] is required.
+ * @property onDownloadCompleted a callback invoked when a download is completed.
+ * @property downloadManager a reference to the [DownloadManager] which is
+ * responsible for performing the downloads.
+ * @property sessionManager a reference to the application's [SessionManager].
+ * @property fragmentManager a reference to a [FragmentManager]. If a fragment
+ * manager is provided, a dialog will be shown before every download.
+ * @property dialog a reference to a [DownloadDialogFragment]. If not provided, an
+ * instance of [SimpleDownloadDialogFragment] will be used.
  */
 class DownloadsFeature(
     private val applicationContext: Context,
@@ -46,7 +49,7 @@ class DownloadsFeature(
 ) : SelectionAwareSessionObserver(sessionManager), LifecycleAwareFeature {
 
     /**
-     * Starts observing any download on the selected session and send it to the [DownloadManager]
+     * Starts observing downloads on the selected session and sends them to the [DownloadManager]
      * to be processed.
      */
     override fun start() {
@@ -58,7 +61,7 @@ class DownloadsFeature(
     }
 
     /**
-     * Stops observing any download on the selected session.
+     * Stops observing downloads on the selected session.
      */
     override fun stop() {
         super.stop()
@@ -66,13 +69,11 @@ class DownloadsFeature(
     }
 
     /**
-     * Notifies to the [DownloadManager] that a new download must be processed.
+     * Notifies the [DownloadManager] that a new download must be processed.
      */
     @SuppressLint("MissingPermission")
     override fun onDownload(session: Session, download: Download): Boolean {
-
         return if (applicationContext.isPermissionGranted(INTERNET, WRITE_EXTERNAL_STORAGE)) {
-
             if (fragmentManager != null) {
                 showDialog(download, session)
                 false
@@ -87,8 +88,8 @@ class DownloadsFeature(
     }
 
     /**
-     * Call this method to notify the feature download that the permissions where granted.
-     * If so it will automatically trigger the pending download.
+     * Notifies the feature that the permissions were granted. It will then
+     * automatically trigger the pending download.
      */
     fun onPermissionsGranted() {
         activeSession?.let { session ->
@@ -96,6 +97,14 @@ class DownloadsFeature(
                 onDownload(session, it)
             }
         }
+    }
+
+    /**
+     * Notifies the feature that the permissions were denied. It will then
+     * clear the pending download.
+     */
+    fun onPermissionsDenied() {
+        activeSession?.download = Consumable.empty()
     }
 
     @SuppressLint("MissingPermission")